+++ core/initializer/test_suite_runner.go
@@ -26,17 +26,6 @@ const (
 )
 
 // =============================== Test Suite Runner =========================================
-type TestSuiteRunner struct {
-	testSuite               testsuite.TestSuite
-	testControllerImageName string
-	testControllerEnvVars   map[string]string
-
-	// The test controller image-specific string representing the log level, that will be passed as-is to the test controller
-	testControllerLogLevel	string
-
-	networkWidthBits uint32
-}
-
 const (
 	// This is the IP address that the first Docker subnet will be doled out from, with subsequent Docker networks doled out with
 	//  increasing IPs corresponding to the NETWORK_WIDTH_BITS
@@ -45,17 +34,39 @@ const (
 	BITS_IN_IP4_ADDR = 32
 )
 
+/*
+An executor to run one or more tests from a given test suite
+ */
+type TestSuiteRunner struct {
+	// The test suite to run tests from
+	testSuite               testsuite.TestSuite
+
+	// The name of the Docker image that should be used to run the test controller container
+	testControllerImageName string
+
+	// Key-value mapping that will be passed as-is to the test controller container on startup in the form of Docker
+	// 	environment variables
+	customTestControllerEnvVars map[string]string
+
+	// A string, meaningful only to the test controller, that represents the log level that the controller container should
+	//	run with
+	testControllerLogLevel	string
+
+	// The number of bits in a test network's subnet mask, such that 2 ^ this_value will be the maximum number of allowed
+	//  services in any given test network
+	networkWidthBits uint32
+}
 
 /*
-Creates a new TestSuiteRunner with the following arguments
+Creates a new TestSuiteRunner with the given parameters.
 
 Args:
-	testSuite: The test suite containing all registered tests
-	testControllerImageName: The name of the Docker image of the test controller that will run the test
+	testSuite: The test suite containing all the user's registered tests
+	testControllerImageName: The name of the Docker image of the test controller that will orchestrate test execution
 	testControllerLogLevel: The string representing the loglevel of the controller (the test suite runner won't be able
 		to parse this, so this should be meaningful to the controller image)
 	networkWidthBits: Each test will get a Docker network with a number of available IP addresses = 2^network_width_bits.
-		This parameter should be set so that all testb
+		This parameter should be set high enough so that each test can fit all the services they want.
  */
 func NewTestSuiteRunner(
 			testSuite testsuite.TestSuite,
@@ -64,11 +75,11 @@ func NewTestSuiteRunner(
 			testControllerEnvVars map[string]string,
 			networkWidthBits uint32) *TestSuiteRunner {
 	return &TestSuiteRunner{
-		testSuite:               testSuite,
-		testControllerImageName: testControllerImageName,
-		testControllerLogLevel:  testControllerLogLevel,
-		testControllerEnvVars:   testControllerEnvVars,
-		networkWidthBits: networkWidthBits,
+		testSuite:                   testSuite,
+		testControllerImageName:     testControllerImageName,
+		testControllerLogLevel:      testControllerLogLevel,
+		customTestControllerEnvVars: testControllerEnvVars,
+		networkWidthBits:            networkWidthBits,
 	}
 }
 
@@ -78,6 +89,11 @@ Runs the tests with the given names and prints the results to STDOUT. If no test
 Args:
 	testNamesToRun: A "set" of test names to run
 	testParallelism: How many tests to run in parallel
+
+Returns:
+	allTestsPassed: True if all tests passed, false otherwise
+	executionErr: An error that will be non-nil if an error occurred that prevented the test from running and/or the result
+		being retrieved. If this is non-nil, the allTestsPassed value is undefined!
  */
 func (runner TestSuiteRunner) RunTests(testNamesToRun map[string]bool, testParallelism uint) (allTestsPassed bool, executionErr error) {
 	allTests := runner.testSuite.GetTests()
@@ -117,7 +133,7 @@ func (runner TestSuiteRunner) RunTests(testNamesToRun map[string]bool, testParal
 		dockerClient,
 		runner.testControllerImageName,
 		runner.testControllerLogLevel,
-		runner.testControllerEnvVars,
+		runner.customTestControllerEnvVars,
 		testParallelism)
 
 	logrus.Infof("Running %v tests with execution ID %v...", len(testsToRun), executionInstanceId.String())